Add a new gitremote package This package is primarily focused on the repository itself, not the checkpoint remote - which has different heuristics.
Entire-Checkpoint: 35c623f9a09d
